test-infra/requirements.txt
google-cloud-storage
google-cloud-spanner
google-cloud-datastream